% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passwordrandom-package.r
\docType{package}
\name{passwordrandom}
\alias{passwordrandom}
\alias{passwordrandom-package}
\title{Use the PasswordRandom.com API}
\description{
Use the PasswordRandom.com API
}
\author{
Bob Rudis (@hrbrmstr)
}
\seealso{
\url{http://www.passwordrandom.com/api}
}

