% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passwordrandom.r
\name{random_chars}
\alias{random_chars}
\title{Generate random characters}
\usage{
random_chars(count = 10, fmt = "ul")
}
\arguments{
\item{count}{number of random characters to generate}

\item{fmt}{\code{ul} == uppercase or lowercase, \code{u} == uppercase,
\code{l} == lowercase}
}
\value{
vector of random characters
}
\description{
Generate random characters
}
\examples{
random_chars(count=100)
}
\seealso{
\url{http://www.passwordrandom.com/api}
}

