% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passwordrandom.r
\name{random_doubles}
\alias{random_doubles}
\title{Generate random doubles}
\usage{
random_doubles(count = 10, min = 0, max = 100)
}
\arguments{
\item{count}{number of random integers to generate}

\item{min}{minimum val of random range}

\item{max}{maximum val of random range}
}
\value{
vector of random doubles
}
\description{
Generate random doubles
}
\examples{
random_doubles(count=100)
}
\seealso{
\url{http://www.passwordrandom.com/api}
}

