% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passwordrandom.r
\name{random_guids}
\alias{random_guids}
\title{Generate random guids}
\usage{
random_guids(count = 10)
}
\arguments{
\item{count}{number of random guids to generate}
}
\value{
vector of random guids
}
\description{
Generate random guids
}
\examples{
random_guids(count=100)
}
\seealso{
\url{http://www.passwordrandom.com/api}
}

