% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/passwordrandom.r
\name{random_passwords}
\alias{random_passwords}
\title{Generate random passwords}
\usage{
random_passwords(count = 10, scheme = "CvvCVN#NNcc")
}
\arguments{
\item{count}{number of random passwords to generate}

\item{complexity}{format string \code{L} - upper case letter;
\code{l} - lower case letter; \code{R} or \code{r} - random case letter;
\code{N} or \code{n} - number (\code{0}-\code{9}); \code{#} - symbol (!@#$.+);
\code{!} - all symbols; \code{C} - upper case consonant letter;
\code{c} - lower case consonant letter; \code{V} - upper case vowel letter;
\code{v} - lower case vowel letter;}
}
\value{
vector of random passwords
}
\description{
Generate random passwords
}
\examples{
random_passwords(count=100)
}
\seealso{
\url{http://www.passwordrandom.com/pronounceable-password-generator}
}

